refactor: 패키지 구조 mvc로 변경
