+++ src/test/java/racewinner/domain/CarTest.java
@@ -1,8 +1,9 @@
-package racewinner;
+package racewinner.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import racewinner.domain.Car;
 
 import static org.assertj.core.api.Assertions.assertThat;
